+++ client/frontend/src/pages/Admin/BookingManagement/ChambalManagement.js
@@ -63,18 +63,21 @@ export default function ChambalManagement() {
 
 
    const HandleDelete = (id) => {
-       axios.delete(`${process.env.REACT_APP_BASE_URL}/admin/bookings/chambal/${id}`, {
-       headers: {
-           'Accept': 'application/json, text/plain, */*',
-           'Content-Type': 'application/json',
-           'Authorization': `Bearer `+localStorage.getItem('accessToken')
-       },
-       }).then(result => {
-           alert.success("Data is deleted");
-           setTimeout(() => {
-           window.location = '/admin/chambal-bookings';
-           }, 1000);
-       })
+    var result = window.confirm("Want to delete?");
+        if (result) {
+            axios.delete(`${process.env.REACT_APP_BASE_URL}/admin/bookings/chambal/${id}`, {
+            headers: {
+                'Accept': 'application/json, text/plain, */*',
+                'Content-Type': 'application/json',
+                'Authorization': `Bearer `+localStorage.getItem('accessToken')
+            },
+            }).then(result => {
+                alert.success("Data is deleted");
+                setTimeout(() => {
+                window.location = '/admin/chambal-bookings';
+                }, 1000);
+            })
+        }
    }
 
    const [filterZone , setFilterZone] = useState('');
