+++ client/frontend/src/pages/Admin/Customers/Customers.js
@@ -61,18 +61,21 @@ export default function Customers() {
 
 
     const HandleDelete = (id , type) => {
-        axios.delete(`${process.env.REACT_APP_BASE_URL}/admin/customers/${type}/${id}`, {
-        headers: {
-            'Accept': 'application/json, text/plain, */*',
-            'Content-Type': 'application/json',
-            'Authorization': `Bearer `+localStorage.getItem('accessToken')
-        },
-        }).then(result => {
-            alert.success("Data is deleted");
-            setTimeout(() => {
-            window.location = '/admin/customers';
-            }, 1000);
-        })
+        var result = window.confirm("Want to delete?");
+        if (result) {
+            axios.delete(`${process.env.REACT_APP_BASE_URL}/admin/customers/${type}/${id}`, {
+            headers: {
+                'Accept': 'application/json, text/plain, */*',
+                'Content-Type': 'application/json',
+                'Authorization': `Bearer `+localStorage.getItem('accessToken')
+            },
+            }).then(result => {
+                alert.success("Data is deleted");
+                setTimeout(() => {
+                window.location = '/admin/customers';
+                }, 1000);
+            })
+        }
     }
 
     const [type , setType] = useState('');
