+++ client/frontend/src/pages/Admin/SafariDates/RanthamboreDates.js
@@ -62,18 +62,21 @@ export default function RanthamboreDates() {
     // kill -9 `sudo lsof -t -i:5003`
 
     const HandleDelete = (id) => {
-        axios.delete(`${process.env.REACT_APP_BASE_URL}/safari/dates/${id}`, {
-          headers: {
-            'Accept': 'application/json, text/plain, */*',
-            'Content-Type': 'application/json',
-            'Authorization': `Bearer `+localStorage.getItem('accessToken')
-          },
-        }).then(result => {
-            alert.success("Data is deleted");
-            setTimeout(() => {
-              window.location = '/admin/ranthambore-dates';
-             }, 1000);
-        })
+        var result = window.confirm("Want to delete?");
+        if (result) {
+            axios.delete(`${process.env.REACT_APP_BASE_URL}/safari/dates/${id}`, {
+            headers: {
+                'Accept': 'application/json, text/plain, */*',
+                'Content-Type': 'application/json',
+                'Authorization': `Bearer `+localStorage.getItem('accessToken')
+            },
+            }).then(result => {
+                alert.success("Data is deleted");
+                setTimeout(() => {
+                window.location = '/admin/ranthambore-dates';
+                }, 1000);
+            })
+        }
     }
 
     const HandleFilter = () => {
